HTTP settings moved to config subpackage
